Added Example Experiments and Experimentor
An environment for automatically executing experiments
with the default W-Model has been added.
The experiments run in parallel and their output is
automatically logged to a folder structure of text log
files.
Furthermore, all runs are started with specific random
generator seeds, which are also stored in the log
files. The experiments are thus 100% reproducible.