error in paper?
In file WModel_Ruggedness the paper, there is a --j which sets j back to
q (which is n). In the paper, it says j=start, which is probably not the
same. This might be an error in our paper, and needs to be investigated.

Thomas Weise and Zijun Wu.
Difficult Features of Combinatorial Optimization Problems and the
Tunable W-Model Benchmark Problem for Simulating them.
In Black Box Discrete Optimization Benchmarking (BB-DOB) Workshop
of Companion Material Proceedings of the Genetic and Evolutionary
Computation Conference (GECCO 2018), July 15th-19th 2018,
Kyoto, Japan, pages 1769-1776, ISBN: 978-1-4503-5764-7. ACM.
doi:10.1145/3205651.3208240